+++ src/transaction/mod.rs
@@ -14,7 +14,9 @@ use std::time::Duration;
 use tracing::debug;
 
 pub use hooks::get_hook;
-pub use install::{install_part, install_part_with_origin, install_parts, install_parts_with_explicit_targets};
+pub use install::{
+    install_part, install_part_with_origin, install_parts, install_parts_with_explicit_targets,
+};
 pub use remove::{
     cascade_remove_list, order_removal_batch, remove_part, remove_part_with_ignored_dependents,
 };